requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
python-dotenv
pgvector
pytest
markdown
pandas 
requests
beautifulsoup4
trafilatura
playwright
notebook
matplotlib
psycopg2-binary 
sqlalchemy
sentence_transformers